add frameBufferScreen  show fps
show frameBuffer texture, it must adjust the camera  for its updownside
versus bitmap on our sight.

+++ src/main/java/com/opengles/book/framework/gl/FPSCounter.java
@@ -14,7 +14,7 @@ public class FPSCounter {
             startTime = System.nanoTime();
         }
     }
-    public long getFps()
+    public int getFps()
     {
         return fps;
     }
